change models to not use Bigint

+++ src/model/Proposal.ts
@@ -3,7 +3,7 @@ import Frame from "./Frame";
 
 export default interface Proposal{
     version: number,
-    genesisVal: BigInt,
+    genesisVal: Uint8Array,
     authors: Author[],
     frames: Frame[]
 }
